fonction supprimer un élément du panier OK

+++ front/js/cart.js
@@ -71,8 +71,8 @@ document.getElementById("totalPrice").innerHTML = totalPrice;
 
 /* gérer la modification et la suppression de produits dans la page Panier : attention à modifier le DOM mais aussi localStorage */
 /* modifier la quantité d'un produit directement sur la page panier si le produit existe déjà */
-let quantityItem = document.getElementsByClassName(".itemQuantity");
-quantityItem.forEach((tag) => {
+/*let quantityItem = document.getElementsByClassName(".itemQuantity");
+quantityItem.addEventListener("click", (tag) => {
   let article = tag.closest("article");
   let id = article.dataset.idKanap;
   let color = article.dataset.colorKanap;
@@ -88,35 +88,39 @@ quantityItem.forEach((tag) => {
     })
   })
 }) 
+*/
+
+  /* sélection des réf de tous les éléments "supprimer" */
+  let deleteItem = document.querySelectorAll(".deleteItem");
+  console.log(deleteItem);
+  if (deleteItem.length != 0){
+    for (let i=0;i<deleteItem.length;i++){
+      deleteItem[i].addEventListener("click", event => {
+        event.preventDefault();
+        console.log(event);
+        /* sélection de l'id du produit à supprimer */
+        let idDeleteItem = productInCart[i].idKanap;
+        let colorDeleteItem = productInCart[i].colorKanap;
+        console.log(idDeleteItem);
+        console.log(colorDeleteItem);
+        /* avec la méthode filter() je sélectionne les éléments à garder et je supprime les éléments où "supprimer" a été cliqué */
+        /* la méthode filter crée et retourne un nouveau tableau contenant tous les éléments du tableau d'origine qui remplissent 
+        une condition déterminée par la fonction callback */
+        productInCart = productInCart.filter(el => el.idKanap !== idDeleteItem || el.colorKanap !== colorDeleteItem);
+        console.log(productInCart); 
+        /* on supprime l'élt du DOM */
+        event.target.closest('article').remove();
+        alert("ce produit a été supprimé de votre panier !");
+        /* on met à jour le localStorage */
+        localStorage.setItem("productToCart",JSON.stringify(productInCart));
+        /* recharger la page pour afficher le nouveau panier */
+        document.location.reload();
+        /* le nombre d'article total et le prix total sont mis à jour automatiquement quand la page se recharge */
+      })
+    }
+  }
 
 
-/* sélection des réf de tous les éléments "supprimer" */
-let deleteItem = document.querySelectorAll(".deleteItem");
-console.log(deleteItem);
-for (let i=0;i<deleteItem.length;i++){
-  deleteItem[i].addEventListener("click", event => {
-    event.preventDefault();
-    console.log(event);
-    /* sélection de l'id du produit à supprimer */
-    let idDeleteItem = productInCart[i].idKanap;
-    let colorDeleteItem = productInCart[i].colorKanap;
-    console.log(idDeleteItem);
-    console.log(colorDeleteItem);
-    /* avec la méthode filter() je sélectionne les éléments à garder et je supprime les éléments où "supprimer" a été cliqué */
-    productInCart = productInCart.filter(el => el.idKanap == idDeleteItem && el.colorKanap == colorDeleteItem);
-    console.log(productInCart);
-    console.log(productInCart.filter(el => el.idKanap == idDeleteItem && el.colorKanap == colorDeleteItem));
-    /* la méthode filter crée et retourne un nouveau tableau contenant tous les éléments du tableau d'origine qui remplissent une condition déterminée par la fonction callback */
-    /* on envoie les éléments non supprimés dans le localStorage */
-    localStorage.setItem("productToCart", JSON.stringify(productInCart));
-    alert("ce produit a été supprimé de votre panier !");
-    /* recharger la page pour afficher le nouveau panier */
-    document.location.reload();
-    /* le nombre d'article total et le prix total sont mis à jour automatiquement quand la page se recharge */
-    
-    })
-
-}
 /* passer la commande : vérifier les données saisies et message d'erreur si nécessaire */
 /*valider le first Name=prénom du formulaire */
 let formFirstName = document.getElementById("firstName");
